Cleanup of file relations on attribute deletion

+++ schema/attribute/attribute.go
@@ -37,23 +37,24 @@ func Del_tx(tx pgx.Tx, id uuid.UUID) error {
 		}
 	}
 
-	// DROP COLUMN removes constraints automatically if there
-	if _, err := tx.Exec(db.Ctx, fmt.Sprintf(`
-		ALTER TABLE "%s"."%s"
-		DROP COLUMN "%s"
-	`, moduleName, relationName, name)); err != nil {
-		return err
+	// delete attribute database entities
+	if schema.IsContentFiles(content) {
+		if err := FileRelationsDelete_tx(tx, id); err != nil {
+			return err
+		}
+	} else {
+		// DROP COLUMN removes constraints if there
+		if _, err := tx.Exec(db.Ctx, fmt.Sprintf(`
+			ALTER TABLE "%s"."%s"
+			DROP COLUMN "%s"
+		`, moduleName, relationName, name)); err != nil {
+			return err
+		}
 	}
-
-	// file attribute content is automatically cleaned up via background task
 
 	// delete attribute reference
-	if _, err := tx.Exec(db.Ctx, `
-		DELETE FROM app.attribute WHERE id = $1
-	`, id); err != nil {
-		return err
-	}
-	return nil
+	_, err = tx.Exec(db.Ctx, `DELETE FROM app.attribute WHERE id = $1`, id)
+	return err
 }
 
 func Get(relationId uuid.UUID) ([]types.Attribute, error) {
@@ -326,7 +327,11 @@ func Set_tx(tx pgx.Tx, relationId uuid.UUID, id uuid.UUID,
 		}
 	} else {
 		// create attribute column (files attribute have no column)
-		if !isFiles {
+		if isFiles {
+			if err := fileRelationsCreate_tx(tx, id, moduleName, relationName); err != nil {
+				return err
+			}
+		} else {
 			// check relationship target if relationship attribute
 			var contentRel string
 			if isRel {
@@ -371,58 +376,6 @@ func Set_tx(tx pgx.Tx, relationId uuid.UUID, id uuid.UUID,
 				ALTER TABLE "%s"."%s" 
 				ADD COLUMN "%s" %s %s %s
 			`, moduleName, relationName, name, columnDef, nullableDef, defaultDef)); err != nil {
-				return err
-			}
-		} else {
-			// create file storage relation
-			tName := schema.GetFilesTableName(id)
-			tNameV := schema.GetFilesTableNameVersions(id)
-			if _, err := tx.Exec(db.Ctx, fmt.Sprintf(`
-				CREATE TABLE instance_file."%s" (
-					id uuid NOT NULL,
-					record_id bigint,
-					name text NOT NULL,
-				    CONSTRAINT "%s_pkey" PRIMARY KEY (id),
-				    CONSTRAINT "%s_record_id_fkey" FOREIGN KEY (record_id)
-				        REFERENCES "%s"."%s" ("%s") MATCH SIMPLE
-				        ON UPDATE SET NULL
-				        ON DELETE SET NULL
-				        DEFERRABLE INITIALLY DEFERRED
-				);
-				
-				CREATE INDEX "fki_%s_record_id_fkey"
-					ON instance_file."%s" USING btree (record_id ASC NULLS LAST);
-				
-				-- file versions
-				CREATE TABLE instance_file."%s" (
-					file_id uuid NOT NULL,
-					version int NOT NULL,
-					login_id integer,
-					hash char(64),
-					size_kb int NOT NULL,
-					date_change bigint NOT NULL,
-				    CONSTRAINT "%s_pkey" PRIMARY KEY (file_id, version),
-				    CONSTRAINT "%s_file_id_fkey" FOREIGN KEY (file_id)
-				        REFERENCES instance_file."%s" (id) MATCH SIMPLE
-				        ON UPDATE CASCADE
-				        ON DELETE CASCADE
-				        DEFERRABLE INITIALLY DEFERRED,
-				    CONSTRAINT "%s_login_id_fkey" FOREIGN KEY (login_id)
-				        REFERENCES instance.login (id) MATCH SIMPLE
-				        ON UPDATE SET NULL
-				        ON DELETE SET NULL
-				        DEFERRABLE INITIALLY DEFERRED
-				);
-				
-				CREATE INDEX "fki_%s_file_id_fkey"
-					ON instance_file."%s" USING btree (file_id ASC NULLS LAST);
-				
-				CREATE INDEX "fki_%s_login_id_fkey"
-					ON instance_file."%s" USING btree (login_id ASC NULLS LAST);
-			`, tName, tName, tName, moduleName, relationName, schema.PkName,
-				tName, tName, tNameV, tNameV, tNameV, tName, tNameV,
-				tNameV, tNameV, tNameV, tNameV)); err != nil {
-
 				return err
 			}
 		}
